Update funscript value type to double to be more precise for devices that support it

+++ Edi.Core/Funscript/CommandLinear.cs
@@ -12,7 +12,7 @@ namespace Edi.Core.Funscript
         public static int SpeedLimit => 400;
 
         #region move to some creation pattern
-        public static CmdLinear GetCommandSpeed(int speed, int value, int initialValue)
+        public static CmdLinear GetCommandSpeed(int speed, double value, double initialValue)
         {
             speed = speed > SpeedLimit ? SpeedLimit : speed;
 
@@ -25,7 +25,7 @@ namespace Edi.Core.Funscript
                 Value = Convert.ToByte(value),
             };
         }
-        public static CmdLinear GetCommand(int millis, int value, int initialValue = 0)
+        public static CmdLinear GetCommand(int millis, double value, double initialValue = 0)
         {
             return new CmdLinear
             {
@@ -34,7 +34,7 @@ namespace Edi.Core.Funscript
                 Value = Convert.ToByte(value)
             };
         }
-        public static CmdLinear GetCommand(uint millis, int value, int initialValue = 0)
+        public static CmdLinear GetCommand(uint millis, double value, double initialValue = 0)
         {
             return new CmdLinear
             {
@@ -69,7 +69,7 @@ namespace Edi.Core.Funscript
             }
             return resul;
         }
-        public static CmdLinear GetCommandMillis(int millis, int value)
+        public static CmdLinear GetCommandMillis(int millis, double value)
         {
 
             return new CmdLinear
@@ -88,10 +88,10 @@ namespace Edi.Core.Funscript
         public int Speed => Millis == 0 ? 0 : Convert.ToInt32(Math.Abs(InitialValue - Value) / (double)Millis * 1000);
 
         public bool Direction => Value > InitialValue;
-        public int Value { get; set; }
+        public double Value { get; set; }
 
-        private int initialValue;
-        public int InitialValue
+        private double initialValue;
+        public double InitialValue
         {
             get => Prev?.Value ?? initialValue;
             set => initialValue = value;
